+++ game_patch/multi/server.cpp
@@ -402,6 +402,18 @@ FunHook<float(rf::Entity*, float, int, int, int)> entity_damage_hook{
             }
         }
 
+        // should entity gib?
+        if (damaged_ep->life < -100.0f && damage_type == 3 &&   // explosive
+            damaged_ep->material == 3 &&                        // flesh
+            rf::game_get_gore_level() >= 2 &&
+            !(damaged_ep->entity_flags & 0x2000000) &&          // custom_corpse (used by snakes and sea creature)
+            !(damaged_ep->entity_flags & 0x1) &&                // dying
+            !(damaged_ep->entity_flags & 0x1000) &&             // in_water
+            !(damaged_ep->entity_flags & 0x2000))               // eye_under_water
+        {
+            damaged_ep->entity_flags |= 0x80;
+        }
+
         float real_damage = entity_damage_hook.call_target(damaged_ep, damage, killer_handle, damage_type, killer_uid);
 
         if (rf::is_server && is_pvp_damage && real_damage > 0.0f) {